Практическая задача 2.2.1 Практическое задание Java pre-project. Hibernate и ORM. Исправленны замечания

+++ 2.2.1. spring_hibernate/src/main/java/hiber/dao/UserDao.java
@@ -6,7 +6,7 @@ import java.util.List;
 
 public interface UserDao {
 
-    User findUserById(int id);
+    User findUserById(Long id);
 
     void add(User user);
